requirement.txt
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
google-generativeai==0.8.4
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
packaging==24.1
pymongo==4.8.0
waitress==3.0.0
reportlab==3.6.13
Werkzeug==3.0.3
azure-storage-blob==12.16.0